refactor : GlobalStyle 파일명 수정
